Split utilities.py into config_file, generic, manifest, and validators modules
