+++ src/xsoar_cli/utilities/generic.py
@@ -1,0 +1,8 @@
+def parse_string_to_dict(input_string: str | None, delimiter: str) -> dict:
+    if not input_string:
+        return {}
+    # Parse a string into a python dictionary
+    pairs = [pair.split("=", 1) for pair in input_string.split(delimiter)]
+    # Filter pairs that have exactly 2 parts (key and value) after splitting by "="
+    valid_pairs = [pair for pair in pairs if len(pair) == 2]  # noqa: PLR2004
+    return {key.strip(): value.strip() for key, value in valid_pairs}
